+++ doc/src/release_notes.rst
@@ -13,6 +13,9 @@ oracledb 1.2.0b1 (TBD)
 Thin Mode Changes
 +++++++++++++++++
 
+#)  Added support for using a wallet for one-way TLS connections, rather than
+    requiring OS recognition of certificates
+    (`issue 65 <https://github.com/oracle/python-oracledb/issues/65>`__).
 #)  Fixed bug when fetching nested cursors with more columns than the parent
     cursor.