+++ doc/src/release_notes.rst
@@ -25,6 +25,8 @@ Thick Mode Changes
 Common Changes
 ++++++++++++++
 
+#)  Added attribute :attr:`DbObjectType.package_name` which contains the name
+    of the package if the type is a PL/SQL type (otherwise, it will be `None`).
 #)  Improved test suite and documentation.